Refactor to create CallContext
CallContext is a new class that encapsulates call information that may be of
relevance by various parties. This class replaces user instance that was passed
around to ExposedAPI subclasses to allow to pass user authentication. With this
refactoring SystemAPI is no longer a special class and Mapper can be simplified
to accept _only_ classes (no more instances) and always instantiate them with
CallContext as argument.