Add SystemAPI.multicall() support.
This change builds on top of CallContext changes earlier to add multicall
support. Multicall is a standard part of (or a very common extenison) XML-RPC
that allows the caller to pack multiple distinct method calls in one request.
The advantage of multicall is to fight latency. Instead of making multiple
calls and paying for the round-trip time the caller can (if possible) pack all
the calls in one multicall request and dispatch them in one go. Each sub-call
returns a distinct return value or error (fault). Fault of any method is not
affecting the ivokation of other methods. This feature is supported by standard
python XML-RPC client.