Fix SystemAPI.muticall() return value.
So I messed up. The implementation worked but returned values incorrectly
(according to the XML-RPC spec). The proper way of returning a value of
subsequent sub-calls is actually trickier and allows to make faults properly
distinguishable from returning structs with identical fields. Normal results
are boxed in one-element array. Faults are returned directly. I added a few
tests to check this behavior and actually tested this with python's
xmlrpclib.MultiCall this time.